+++ cmd/tfproviderlint/tfproviderlint.go
@@ -32,6 +32,7 @@ import (
 	"github.com/bflad/tfproviderlint/passes/S013"
 	"github.com/bflad/tfproviderlint/passes/S014"
 	"github.com/bflad/tfproviderlint/passes/S015"
+	"github.com/bflad/tfproviderlint/passes/S016"
 )
 
 func main() {
@@ -59,5 +60,6 @@ func main() {
 		S013.Analyzer,
 		S014.Analyzer,
 		S015.Analyzer,
+		S016.Analyzer,
 	)
 }