+++ passes/S016/testdata/src/a/alias.go
@@ -1,0 +1,31 @@
+package a
+
+import (
+	s "github.com/hashicorp/terraform/helper/schema"
+)
+
+func falias() {
+	_ = s.Schema{ // want "schema Set should only be included for TypeSet"
+		Set:  set,
+		Type: s.TypeList,
+	}
+
+	_ = s.Schema{
+		Set:  set,
+		Type: s.TypeSet,
+	}
+
+	_ = map[string]*s.Schema{
+		"name": { // want "schema Set should only be included for TypeSet"
+			Set:  set,
+			Type: s.TypeList,
+		},
+	}
+
+	_ = map[string]*s.Schema{
+		"name": {
+			Set:  set,
+			Type: s.TypeSet,
+		},
+	}
+}