+++ passes/S016/testdata/src/a/comment_ignore.go
@@ -1,0 +1,21 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform/helper/schema"
+)
+
+func fcommentignore() {
+	//lintignore:S016
+	_ = schema.Schema{
+		Set:  set,
+		Type: schema.TypeList,
+	}
+
+	//lintignore:S016
+	_ = map[string]*schema.Schema{
+		"name": {
+			Set:  set,
+			Type: schema.TypeList,
+		},
+	}
+}