+++ passes/S016/testdata/src/a/outside_package.go
@@ -1,0 +1,19 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Schema{
+		Set:  func() {},
+		Type: schema.TypeList,
+	}
+
+	_ = map[string]*schema.Schema{
+		"name": {
+			Set:  func() {},
+			Type: schema.TypeList,
+		},
+	}
+}